move type advantage related methods to new class
