+++ src/battle/Battle.java
@@ -30,7 +30,6 @@ import battle.effect.generic.Weather;
 import item.ItemNamesies;
 import main.Game;
 import main.Global;
-import type.Type;
 import map.TerrainType;
 import message.MessageUpdate;
 import message.MessageUpdate.Update;
@@ -47,6 +46,7 @@ import trainer.Team;
 import trainer.Trainer;
 import trainer.Trainer.Action;
 import trainer.WildPokemon;
+import type.TypeAdvantage;
 import util.PokeString;
 import util.RandomUtils;
 
@@ -556,8 +556,8 @@ public class Battle {
 		int attackStat = Stat.getStat(attacking, me, o, this);
 		int defenseStat = Stat.getStat(defending, o, me, this);
 		
-		double stab = Type.getSTAB(this, me);
-		double adv = Type.getAdvantage(me, o, this);
+		double stab = TypeAdvantage.getSTAB(this, me);
+		double adv = TypeAdvantage.getAdvantage(me, o, this);
 		
 		int damage = (int)Math.ceil(((((2*level/5.0 + 2)*attackStat*power/defenseStat)/50.0) + 2)*stab*adv*random/100.0);
