+++ src/main/TeamPlanner.java
@@ -8,6 +8,7 @@ import pokemon.PokemonNamesies;
 import pokemon.Stat;
 import pokemon.ability.AbilityNamesies;
 import type.Type;
+import type.TypeAdvantage;
 import util.FileIO;
 import util.StringUtils;
 
@@ -304,8 +305,8 @@ public class TeamPlanner {
 						int first = firstType.getIndex();
 						int second = secondType.getIndex();
 						
-						double firstAdvantage = Type.getBasicAdvantage(attackType, firstType);
-						double secondAdvantage = Type.getBasicAdvantage(attackType, secondType);
+						double firstAdvantage = TypeAdvantage.getBasicAdvantage(attackType, firstType);
+						double secondAdvantage = TypeAdvantage.getBasicAdvantage(attackType, secondType);
 						
 						double advantage = firstAdvantage*secondAdvantage;
